Move get_url2 to c2cgeoportal_commons.lib
