+++ commons/requirements.txt
@@ -1,6 +1,7 @@
 c2c.template
 GeoAlchemy2
 papyrus
+pyramid
 sqlalchemy
 transaction
 zope.event
